Update of AlphaFS to latest version

+++ AlphaFS/Filesystem/Directory Class/Directory.Delete.cs
@@ -20,6 +20,7 @@
  */
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.IO;
@@ -44,7 +45,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void Delete(string path)
       {
-         DeleteDirectoryCore(null, null, path, false, false, true, false, PathFormat.RelativePath);
+         DeleteDirectoryCore(null, null, path, false, false, false, PathFormat.RelativePath);
       }
 
       /// <summary>Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
@@ -60,7 +61,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void Delete(string path, bool recursive)
       {
-         DeleteDirectoryCore(null, null, path, recursive, false, !recursive, false, PathFormat.RelativePath);
+         DeleteDirectoryCore(null, null, path, recursive, false, false, PathFormat.RelativePath);
       }
 
       #endregion // .NET
@@ -78,9 +79,8 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void Delete(string path, PathFormat pathFormat)
       {
-         DeleteDirectoryCore(null, null, path, false, false, true, false, pathFormat);
+         DeleteDirectoryCore(null, null, path, false, false, false, pathFormat);
       }
-
 
 
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
@@ -97,9 +97,8 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void Delete(string path, bool recursive, PathFormat pathFormat)
       {
-         DeleteDirectoryCore(null, null, path, recursive, false, !recursive, false, pathFormat);
+         DeleteDirectoryCore(null, null, path, recursive, false, false, pathFormat);
       }
-
 
 
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
@@ -116,7 +115,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void Delete(string path, bool recursive, bool ignoreReadOnly)
       {
-         DeleteDirectoryCore(null, null, path, recursive, ignoreReadOnly, !recursive, false, PathFormat.RelativePath);
+         DeleteDirectoryCore(null, null, path, recursive, ignoreReadOnly, false, PathFormat.RelativePath);
       }
 
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
@@ -134,8 +133,9 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void Delete(string path, bool recursive, bool ignoreReadOnly, PathFormat pathFormat)
       {
-         DeleteDirectoryCore(null, null, path, recursive, ignoreReadOnly, !recursive, false, pathFormat);
+         DeleteDirectoryCore(null, null, path, recursive, ignoreReadOnly, false, pathFormat);
       }
+
 
       #region Transactional
 
@@ -152,7 +152,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void DeleteTransacted(KernelTransaction transaction, string path)
       {
-         DeleteDirectoryCore(null, transaction, path, false, false, true, false, PathFormat.RelativePath);
+         DeleteDirectoryCore(null, transaction, path, false, false, false, PathFormat.RelativePath);
       }
 
       /// <summary>[AlphaFS] Deletes an empty directory from a specified path.</summary>
@@ -169,11 +169,10 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void DeleteTransacted(KernelTransaction transaction, string path, PathFormat pathFormat)
       {
-         DeleteDirectoryCore(null, transaction, path, false, false, true, false, pathFormat);
+         DeleteDirectoryCore(null, transaction, path, false, false, false, pathFormat);
       }
 
-
-
+      
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
       /// <exception cref="ArgumentException"/>
       /// <exception cref="ArgumentNullException"/>
@@ -188,7 +187,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void DeleteTransacted(KernelTransaction transaction, string path, bool recursive)
       {
-         DeleteDirectoryCore(null, transaction, path, recursive, false, !recursive, false, PathFormat.RelativePath);
+         DeleteDirectoryCore(null, transaction, path, recursive, false, false, PathFormat.RelativePath);
       }
 
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
@@ -206,10 +205,9 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void DeleteTransacted(KernelTransaction transaction, string path, bool recursive, PathFormat pathFormat)
       {
-         DeleteDirectoryCore(null, transaction, path, recursive, false, !recursive, false, pathFormat);
+         DeleteDirectoryCore(null, transaction, path, recursive, false, false, pathFormat);
       }
-
-
+      
 
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
       /// <exception cref="ArgumentException"/>
@@ -226,7 +224,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void DeleteTransacted(KernelTransaction transaction, string path, bool recursive, bool ignoreReadOnly)
       {
-         DeleteDirectoryCore(null, transaction, path, recursive, ignoreReadOnly, !recursive, false, PathFormat.RelativePath);
+         DeleteDirectoryCore(null, transaction, path, recursive, ignoreReadOnly, false, PathFormat.RelativePath);
       }
 
       /// <summary>[AlphaFS] Deletes the specified directory and, if indicated, any subdirectories in the directory.</summary>
@@ -245,10 +243,11 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static void DeleteTransacted(KernelTransaction transaction, string path, bool recursive, bool ignoreReadOnly, PathFormat pathFormat)
       {
-         DeleteDirectoryCore(null, transaction, path, recursive, ignoreReadOnly, !recursive, false, pathFormat);
+         DeleteDirectoryCore(null, transaction, path, recursive, ignoreReadOnly, false, pathFormat);
       }
 
       #endregion // Transactional
+
 
       #region Internal Methods
 
@@ -261,76 +260,81 @@ namespace Alphaleonis.Win32.Filesystem
       /// <exception cref="NotSupportedException"/>
       /// <exception cref="UnauthorizedAccessException"/>
       /// <exception cref="DirectoryReadOnlyException"/>
-      /// <param name="fileSystemEntryInfo">A FileSystemEntryInfo instance. Use either <paramref name="fileSystemEntryInfo"/> or <paramref name="path"/>, not both.</param>
+      /// <param name="fsEntryInfo">A FileSystemEntryInfo instance. Use either <paramref name="fsEntryInfo"/> or <paramref name="path"/>, not both.</param>
       /// <param name="transaction">The transaction.</param>
-      /// <param name="path">The name of the directory to remove. Use either <paramref name="path"/> or <paramref name="fileSystemEntryInfo"/>, not both.</param>
+      /// <param name="path">The name of the directory to remove. Use either <paramref name="path"/> or <paramref name="fsEntryInfo"/>, not both.</param>
       /// <param name="recursive"><see langword="true"/> to remove all files and subdirectories recursively; <see langword="false"/> otherwise only the top level empty directory.</param>
       /// <param name="ignoreReadOnly"><see langword="true"/> overrides read only attribute of files and directories.</param>
-      /// <param name="requireEmpty"><see langword="true"/> requires the directory must be empty.</param>
-      /// <param name="continueOnNotExist"><see langword="true"/> does not throw an Exception when the file system object does not exist.</param>
+      /// <param name="continueOnNotFound">When <see langword="true"/> does not throw an Exception when the directory does not exist.</param>
       /// <param name="pathFormat">Indicates the format of the path parameter(s).</param>
       [SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity")]
       [SecurityCritical]
-      internal static void DeleteDirectoryCore(FileSystemEntryInfo fileSystemEntryInfo, KernelTransaction transaction, string path, bool recursive, bool ignoreReadOnly, bool requireEmpty, bool continueOnNotExist, PathFormat pathFormat)
+      internal static void DeleteDirectoryCore(FileSystemEntryInfo fsEntryInfo, KernelTransaction transaction, string path, bool recursive, bool ignoreReadOnly, bool continueOnNotFound, PathFormat pathFormat)
       {
          #region Setup
 
-         if (pathFormat == PathFormat.RelativePath)
-            Path.CheckSupportedPathFormat(path, true, true);
-
-         if (fileSystemEntryInfo == null)
+         if (fsEntryInfo == null)
          {
             // MSDN: .NET 3.5+: DirectoryNotFoundException:
             // Path does not exist or could not be found.
             // Path refers to a file instead of a directory.
             // The specified path is invalid (for example, it is on an unmapped drive). 
 
-            fileSystemEntryInfo = File.GetFileSystemEntryInfoCore(true, transaction, Path.GetExtendedLengthPathCore(transaction, path, pathFormat, GetFullPathOptions.TrimEnd | GetFullPathOptions.RemoveTrailingDirectorySeparator), continueOnNotExist, pathFormat);
+            if (pathFormat == PathFormat.RelativePath)
+               Path.CheckSupportedPathFormat(path, true, true);
+
+            fsEntryInfo = File.GetFileSystemEntryInfoCore(true, transaction, Path.GetExtendedLengthPathCore(transaction, path, pathFormat, GetFullPathOptions.TrimEnd | GetFullPathOptions.RemoveTrailingDirectorySeparator), continueOnNotFound, pathFormat);
+
+            if (fsEntryInfo == null)
+               return;
          }
-
-         if (fileSystemEntryInfo == null)
-            return;
-
-         string pathLp = fileSystemEntryInfo.LongFullPath;
 
          #endregion // Setup
 
+
          // Do not follow mount points nor symbolic links, but do delete the reparse point itself.
-
          // If directory is reparse point, disable recursion.
-         if (recursive && fileSystemEntryInfo.IsReparsePoint)
-            recursive = false;
 
-
-         // Check to see if this is a mount point, and unmount it.
-         if (fileSystemEntryInfo.IsMountPoint)
+         if (recursive && !fsEntryInfo.IsReparsePoint)
          {
-            int lastError = Volume.DeleteVolumeMountPointCore(pathLp, true);
+            var dirs = new Stack<string>(1000);
 
-            if (lastError != Win32Errors.ERROR_SUCCESS && lastError != Win32Errors.ERROR_PATH_NOT_FOUND)
-               NativeError.ThrowException(lastError, pathLp);
-
-            // Now it is safe to delete the actual directory.
-         }
-
-
-         if (recursive)
-         {
-            // Enumerate all file system objects.
-            foreach (var fsei in EnumerateFileSystemEntryInfosCore<FileSystemEntryInfo>(transaction, pathLp, Path.WildcardStarMatchAll, DirectoryEnumerationOptions.FilesAndFolders, PathFormat.LongFullPath))
+            foreach (var fsei in EnumerateFileSystemEntryInfosCore<FileSystemEntryInfo>(transaction, fsEntryInfo.LongFullPath, Path.WildcardStarMatchAll, DirectoryEnumerationOptions.FilesAndFolders | DirectoryEnumerationOptions.Recursive, PathFormat.LongFullPath))
             {
                if (fsei.IsDirectory)
-                  DeleteDirectoryCore(fsei, transaction, null, true, ignoreReadOnly, requireEmpty, true, PathFormat.LongFullPath);
+               {
+                  // Check to see if this is a mount point, and unmount it.
+                  // Now it is safe to delete the actual directory.
+                  if (fsei.IsMountPoint)
+                     Volume.DeleteVolumeMountPointCore(fsei.LongFullPath, false);
+
+                  dirs.Push(fsei.LongFullPath);
+               }
+
                else
                   File.DeleteFileCore(transaction, fsei.LongFullPath, ignoreReadOnly, PathFormat.LongFullPath);
             }
+
+
+            while (dirs.Count > 0)
+               DeleteDirectoryCore(transaction, dirs.Pop(), ignoreReadOnly, continueOnNotFound);
          }
 
-         #region Remove
 
+         // Check to see if this is a mount point, and unmount it.
+         // Now it is safe to delete the actual directory.
+         if (fsEntryInfo.IsMountPoint)
+            Volume.DeleteVolumeMountPointCore(fsEntryInfo.LongFullPath, false);
+
+         DeleteDirectoryCore(transaction, fsEntryInfo.LongFullPath, ignoreReadOnly, continueOnNotFound);
+      }
+      
+
+      private static void DeleteDirectoryCore(KernelTransaction transaction, string pathLp, bool ignoreReadOnly, bool continueOnNotFound)
+      {
          startRemoveDirectory:
 
-         if (!(transaction == null || !NativeMethods.IsAtLeastWindowsVista
+         var success = transaction == null || !NativeMethods.IsAtLeastWindowsVista
 
             // RemoveDirectory() / RemoveDirectoryTransacted()
             // In the ANSI version of this function, the name is limited to MAX_PATH characters.
@@ -340,17 +344,17 @@ namespace Alphaleonis.Win32.Filesystem
             // RemoveDirectory on a symbolic link will remove the link itself.
 
             ? NativeMethods.RemoveDirectory(pathLp)
-            : NativeMethods.RemoveDirectoryTransacted(pathLp, transaction.SafeHandle)))
+            : NativeMethods.RemoveDirectoryTransacted(pathLp, transaction.SafeHandle);
+
+
+         var lastError = Marshal.GetLastWin32Error();
+         if (!success)
          {
-            int lastError = Marshal.GetLastWin32Error();
             switch ((uint) lastError)
             {
                case Win32Errors.ERROR_DIR_NOT_EMPTY:
-                  if (requireEmpty)
-                     // MSDN: .NET 3.5+: IOException: The directory specified by path is not an empty directory. 
-                     throw new DirectoryNotEmptyException(pathLp);
-
-                  goto startRemoveDirectory;
+                  // MSDN: .NET 3.5+: IOException: The directory specified by path is not an empty directory. 
+                  throw new DirectoryNotEmptyException(pathLp);
 
 
                case Win32Errors.ERROR_DIRECTORY:
@@ -361,18 +365,20 @@ namespace Alphaleonis.Win32.Filesystem
 
 
                case Win32Errors.ERROR_PATH_NOT_FOUND:
-                  if (continueOnNotExist)
+                  if (continueOnNotFound)
                      return;
                   break;
 
+               
                case Win32Errors.ERROR_SHARING_VIOLATION:
                   // MSDN: .NET 3.5+: IOException: The directory is being used by another process or there is an open handle on the directory.
                   NativeError.ThrowException(lastError, pathLp);
                   break;
 
+               
                case Win32Errors.ERROR_ACCESS_DENIED:
                   var data = new NativeMethods.WIN32_FILE_ATTRIBUTE_DATA();
-                  int dataInitialised = File.FillAttributeInfoCore(transaction, pathLp, ref data, false, true);
+                  var dataInitialised = File.FillAttributeInfoCore(transaction, pathLp, ref data, false, true);
 
                   if (data.dwFileAttributes != (FileAttributes) (-1))
                   {
@@ -383,7 +389,7 @@ namespace Alphaleonis.Win32.Filesystem
                         if (ignoreReadOnly)
                         {
                            // Reset directory attributes.
-                           File.SetAttributesCore(true, transaction, pathLp, FileAttributes.Normal, true, PathFormat.LongFullPath);
+                           File.SetAttributesCore(true, transaction, pathLp, FileAttributes.Normal, PathFormat.LongFullPath);
                            goto startRemoveDirectory;
                         }
 
@@ -392,8 +398,9 @@ namespace Alphaleonis.Win32.Filesystem
                      }
                   }
 
+
+                  // MSDN: .NET 3.5+: UnauthorizedAccessException: The caller does not have the required permission.
                   if (dataInitialised == Win32Errors.ERROR_SUCCESS)
-                     // MSDN: .NET 3.5+: UnauthorizedAccessException: The caller does not have the required permission.
                      NativeError.ThrowException(lastError, pathLp);
 
                   break;
@@ -408,8 +415,6 @@ namespace Alphaleonis.Win32.Filesystem
 
             NativeError.ThrowException(lastError, pathLp);
          }
-
-         #endregion // Remove
       }
 
       #endregion // Internal Methods
